fix require cycles warning by moving the registerInterceptor function
