+++ src/services/authCredentials/Providers/AuthCredentialsProvider.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect} from 'react';
 import {createContext, useState} from 'react';
 
-import {registerInterceptor} from '@api';
+import {api} from '@api';
 
 import {authService} from '../../../domain/Auth/authService';
 import {AuthCredentials} from '../../../domain/Auth/authTypes';
@@ -74,4 +74,46 @@ export function AuthCredentialsProvider({children}: React.PropsWithChildren) {
       {children}
     </AuthCredentialsContext.Provider>
   );
+}
+
+type InterceptorProps = {
+  authCredentials: AuthCredentials | null;
+  removeCredentials: () => Promise<void>;
+  saveCredentials: (ac: AuthCredentials) => Promise<void>;
+};
+
+export function registerInterceptor({
+  authCredentials,
+  removeCredentials,
+  saveCredentials,
+}: InterceptorProps) {
+  const interceptor = api.interceptors.response.use(
+    response => response,
+    async responseError => {
+      const failedRequest = responseError.config;
+      const hasNotRefreshToken = !authCredentials?.refreshToken;
+      const isRefreshTokenRequest =
+        authService.isRefreshTokenRequest(failedRequest);
+
+      if (responseError.response.status === 401) {
+        if (hasNotRefreshToken || isRefreshTokenRequest || failedRequest.sent) {
+          removeCredentials();
+          return Promise.reject(responseError);
+        }
+
+        failedRequest.sent = true;
+
+        const newAuthCredentials = await authService.authenticateByRefreshToken(
+          authCredentials.refreshToken,
+        );
+        saveCredentials(newAuthCredentials);
+
+        failedRequest.headers.Authorization = `Bearer ${newAuthCredentials.token}`;
+        return api(failedRequest);
+      }
+      return Promise.reject(responseError);
+    },
+  );
+  //remove listen quando o componente é desmontado
+  return () => api.interceptors.response.eject(interceptor);
 }
